feat: EC hier bodies, bucket deletion, seal enforcement, replica support
- Hierarchical bucket bodies now use the same inline-vs-EC path as flat
  objects; fix silent truncation of hier slices beyond HEAD_EMBEDDED_SLICES
- Bucket deletion data flow (PD tombstone → MetaNode DeleteRange → GcRound)
  and GcRound round-bookkeeping at PD
- Seal invariant 4 enforced as hard fence in apply_rebind; InspectRound
  segment-resume for coordinator crash recovery
- Replica support as CodeMode(data=1) via three wirings: per-chunk code
  resolution, seed-then-backfill reads, and explicit write_quorum field
- Presigned URL generation for console object-download 302 redirect;
  /metrics registration for gateway/meta/pd/store; admin HTTP endpoints;
  cross-partition LIST merge; hier rename/rmdir RPCs wired to state machine

+++ crates/epoch-store/src/disk.rs
@@ -269,8 +269,11 @@ fn space_stats_impl(root: &Path) -> (u64, u64, u64) {
     if unsafe { libc::statvfs(path.as_ptr(), &mut stat) } != 0 {
         return (0, 0, 0);
     }
-    let total = u64::from(stat.f_blocks) * stat.f_frsize;
-    let free = u64::from(stat.f_bavail) * stat.f_frsize;
+    // `f_blocks`/`f_bavail`/`f_frsize` are all `fsblkcnt_t`/`c_ulong` = u64 on
+    // the supported target (x86_64 Linux, 99 v0.12 Linux-first), so the products
+    // need no widening conversion.
+    let total = stat.f_blocks * stat.f_frsize;
+    let free = stat.f_bavail * stat.f_frsize;
     (total, free, total.saturating_sub(free))
 }
